refactor(agent-view): share collapsible-toggle helpers via leaf module
The expand/collapse wiring for collapsible rows was factored into
`wireCollapsibleToggle` / `setCollapsibleExpanded` inside
`agent-view-tool-display.ts`, but those helpers were module-private.
`agent-view-messages.ts`'s `renderReasoningSection` re-implemented the
identical behavior by hand on the same `gemini-tool-row-*` class family
(display toggle, chevron swap, `*-expanded` class, `aria-expanded`, and a
duplicate Enter/Space keydown handler).

Move the two helpers and the `CollapsibleRefs` interface into a new leaf
module `src/ui/agent-view/collapsible.ts`, import them from
`agent-view-tool-display.ts`, and call `wireCollapsibleToggle` from
`renderReasoningSection` instead of the hand-rolled copy. Pure code motion
plus reuse; no behavior change.
